AutoTurnStationary
Based off AutoMoveToDistance, this command will be able to turn in place using the gyro and odometry. Also created new program: UT_AutoTurnStationary. This is a mockito test. Using this i hope to be able to test the programs without the robot, to make sure im receiving the correct values. So far it just tests the initialization stage and is suppose to return false, meaning it has initialized and the overall program isn't finished.